mmpm/requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
audioread==3.0.1
beautifulsoup4==4.12.3
black==21.4b2
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
davis2017==0.1.0
decorator==4.4.2
diffusers==0.31.0
dnspython==2.6.1
docker-pycreds==0.4.0
easydict==1.13
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-auth==2.35.0
google-auth-oauthlib==1.0.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.67.0
h11==0.14.0
h5py==3.11.0
hickle==5.0.3
httpcore==1.0.7
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.3
jpeg4py==0.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
mdurl==0.1.2
mistune==3.0.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
motmetrics==1.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.1
numba==0.58.1
numpy==1.24.3
nvidia-ml-py==11.525.112
nvitop==1.1.2
oauthlib==3.2.2
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.1
pandas==2.0.3
pbr==6.0.0
pillow==10.3.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
progressbar2==4.4.2
protobuf==5.27.3
psutil==5.9.5
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydot==2.0.0
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
python-utils==3.8.2
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruff==0.5.0
safetensors==0.4.5
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==69.5.1
Shapely==1.7.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
starlette==0.41.3
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
testresources==2.0.1
thinplate==1.0.0
tifffile==2023.7.10
tikzplotlib==0.10.1
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tornado==6.4.1
tqdm==4.67.1
transformers==4.46.3
triton==2.1.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.32.1
uvloop==0.19.0
visdom==0.2.4
wandb==0.17.5
watchdog==4.0.1
watchfiles==0.22.0
webcolors==24.6.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
wheel==0.43.0
xmltodict==0.13.0
zipp==3.20.2
